JuliaMBD

Documentation for JuliaMBD.

Base.:=>

JuliaMBD.Auto

JuliaMBD.InPort

JuliaMBD.Line

JuliaMBD.OutPort

JuliaMBD.SymbolicValue

JuliaMBD.UndefBlock

JuliaMBD.expr_refvalue

JuliaMBD.expr_refvalue

JuliaMBD.expr_refvalue

JuliaMBD.expr_setvalue

JuliaMBD.expr_setvalue

JuliaMBD.expr_setvalue

JuliaMBD.get_default_inport

JuliaMBD.get_default_inport

JuliaMBD.get_default_inport

JuliaMBD.get_line

JuliaMBD.get_lines

JuliaMBD.get_name

JuliaMBD.get_name

JuliaMBD.get_parent

JuliaMBD.get_source

JuliaMBD.get_target

JuliaMBD.get_var

JuliaMBD.get_var

JuliaMBD.set_line!

JuliaMBD.set_line!

Base.:=> — Method

Base.:(=>)(out, in)

The method to create a line connecting two componets. When it uses with blocks, get_default_inport/get_default_outport are called in the method. The returned value is the out

JuliaMBD.Auto — Type

 Type Auto

Auto means the variable type is determined automatically from the code

JuliaMBD.InPort — Type

InPort <: AbstractInPort

The mutable struct for InPort. InPort corresponds to an argument of the function of block. The members are

var: A symbol of port. SymbolicValue is generic for the parametric types SymbolicValue{Tv}

parent: A block involves the port itself.

line: An instance of Line that connects from other outport to the inport

Examples

jldoctest julia> x = InPort() # name is determined by gensym, Type is Auto julia> x = InPort(:x) # name is :x, Type is Auto julia> x = InPort(Float64) # name is determined by gensym, Type is Float64 julia> x = InPort(:x, Float64) # name is :x, Type is Float64`

JuliaMBD.Line — Type

Line

A mutable struct of Line to connect from outport to inport. The members are

var: A symbol of line. SymbolicValue is generic for the parametric types SymbolicValue{Tv}. But now, the type is fixed as Auto.

source: An instance of outport

dest: An instance of inport

JuliaMBD.OutPort — Type

OutPort <: AbstractOutPort

The mutable struct for OutPort. OutPort corresponds to a returend value from the function of block. The members are

var: A symbol of port. SymbolicValue is generic for the parametric types SymbolicValue{Tv}

parent: A block involves the port itself.

lines: A vector of instances of Line that connects from the outport to others.

Examples

jldoctest julia> x = OutPort() # name is determined by gensym, Type is Auto julia> x = OutPort(:x) # name is :x, Type is Auto julia> x = OutPort(Float64) # name is determined by gensym, Type is Float64 julia> x = OutPort(:x, Float64) # name is :x, Type is Float64`

JuliaMBD.SymbolicValue — Type

SymbolicValue{Tv}

The type expresses a symbolic parameter

JuliaMBD.UndefBlock — Type

 UndefBlock

A struct of UndefiedBlock

JuliaMBD.expr_refvalue — Method

expr_refvalue(x)

Create Expr to refer the line.

JuliaMBD.expr_refvalue — Method

expr_refvalue(x)

Create Expr to refer the port.

JuliaMBD.expr_refvalue — Method

expr_refvalue(x)

Create Expr to refer the symbolic value.

JuliaMBD.expr_setvalue — Method

expr_setvalue(x, expr)

Create Expr to set x = expr.

JuliaMBD.expr_setvalue — Method

expr_setvalue(x, expr)

Create Expr to set x = expr.

JuliaMBD.expr_setvalue — Method

expr_setvalue(x, expr)

Create Expr to set x = expr.

JuliaMBD.get_default_inport — Method

getdefaultinport(p::AbstractInPort) getdefaultoutport(p::AbstractInPort)

Get an instance of default in/out port.

JuliaMBD.get_default_inport — Method

getdefaultinport(p::AbstractOutPort) getdefaultoutport(p::AbstractOutPort)

Get an instance of default in/out port.

JuliaMBD.get_default_inport — Method

get_default_inport(b::UndefBlock)
get_default_outport(b::UndefBlock)

Return a nothing as an instance of default in/out port.

JuliaMBD.get_line — Method

get_line(x::AbstractInPort)

Get an input line of InPort.

JuliaMBD.get_lines — Method

get_lines(x::AbstractOutPort)

Get output lines from OutPort.

JuliaMBD.get_name — Method

get_name(x::AbstractPort)

Get a name (Symbol).

JuliaMBD.get_name — Method

get_name(x::SymbolicValue)

Get a name (Symbol).

JuliaMBD.get_parent — Method

get_parent(x::AbstractPort)

Get a parent block.

JuliaMBD.get_source — Method

get_source(x::AbstractLine)

Get a source of line.

JuliaMBD.get_target — Method

get_target(x::AbstractLine)

Get a target of line.

JuliaMBD.get_var — Method

get_var(x::AbstractLine)

Get a symbolicvar.

JuliaMBD.get_var — Method

get_var(x::AbstractPort)

Get a symbolicvar.

JuliaMBD.set_line! — Method

set_line!(x::AbstractInPort, l::AbstractLine)

Set an input line to InPort.

JuliaMBD.set_line! — Method

set_line!(x::AbstractOutPort, l::AbstractLine)

Set an output line to OutPort.