JuliaMBD

Documentation for JuliaMBD.

Base.:=>

JuliaMBD.Auto

JuliaMBD.BlockDefinition

JuliaMBD.InPort

JuliaMBD.Line

JuliaMBD.OutPort

JuliaMBD.SymbolicValue

JuliaMBD.UndefBlock

JuliaMBD.UndefLine

JuliaMBD.Gain

JuliaMBD._addinport!

JuliaMBD._addoutport!

JuliaMBD._compile_constructor

JuliaMBD._compile_function

JuliaMBD.addblock!

JuliaMBD.addparameter!

JuliaMBD.expr

JuliaMBD.expr_call

JuliaMBD.expr_defvalue

JuliaMBD.expr_func

JuliaMBD.expr_refvalue

JuliaMBD.expr_refvalue

JuliaMBD.expr_refvalue

JuliaMBD.expr_setvalue

JuliaMBD.expr_setvalue

JuliaMBD.expr_setvalue

JuliaMBD.get_default_inport

JuliaMBD.get_default_inport

JuliaMBD.get_default_inport

JuliaMBD.get_default_inport

JuliaMBD.get_default_outport

JuliaMBD.get_inports

JuliaMBD.get_line

JuliaMBD.get_lines

JuliaMBD.get_name

JuliaMBD.get_name

JuliaMBD.get_outports

JuliaMBD.get_params

JuliaMBD.get_parent

JuliaMBD.get_source

JuliaMBD.get_target

JuliaMBD.get_type

JuliaMBD.get_var

JuliaMBD.get_var

JuliaMBD.next

JuliaMBD.prev

JuliaMBD.set_line!

JuliaMBD.set_line!

JuliaMBD.set_parent!

JuliaMBD.tsort

Base.:=> — Method

Base.:(=>)(out, in)

The method to create a line connecting two componets. When it uses with blocks, get_default_inport/get_default_outport are called in the method. The returned value is the out

JuliaMBD.Auto — Type

 Type Auto

Auto means the variable type is determined automatically from the code

JuliaMBD.BlockDefinition — Type

BlockDefinition

A struct to create the following functions:

constructor

expr_call

function

JuliaMBD.InPort — Type

InPort <: AbstractInPort

The mutable struct for InPort. InPort corresponds to an argument of the function of block. The members are

var: A symbol of port. SymbolicValue is generic for the parametric types SymbolicValue{Tv}

parent: A block involves the port itself.

line: An instance of Line that connects from other outport to the inport

Examples

jldoctest julia> x = InPort() # name is determined by gensym, Type is Auto julia> x = InPort(:x) # name is :x, Type is Auto julia> x = InPort(Float64) # name is determined by gensym, Type is Float64 julia> x = InPort(:x, Float64) # name is :x, Type is Float64`

JuliaMBD.Line — Type

Line

A mutable struct of Line to connect from outport to inport. The members are

var: A symbol of line. SymbolicValue is generic for the parametric types SymbolicValue{Tv}. But now, the type is fixed as Auto.

source: An instance of outport

dest: An instance of inport

JuliaMBD.OutPort — Type

OutPort <: AbstractOutPort

The mutable struct for OutPort. OutPort corresponds to a returend value from the function of block. The members are

var: A symbol of port. SymbolicValue is generic for the parametric types SymbolicValue{Tv}

parent: A block involves the port itself.

lines: A vector of instances of Line that connects from the outport to others.

Examples

jldoctest julia> x = OutPort() # name is determined by gensym, Type is Auto julia> x = OutPort(:x) # name is :x, Type is Auto julia> x = OutPort(Float64) # name is determined by gensym, Type is Float64 julia> x = OutPort(:x, Float64) # name is :x, Type is Float64`

JuliaMBD.SymbolicValue — Type

SymbolicValue{Tv}

The type expresses a symbolic parameter

JuliaMBD.UndefBlock — Type

 UndefBlock

A struct of UndefiedBlock

JuliaMBD.UndefLine — Type

UndefLine

A type meaning Undef

JuliaMBD.Gain — Method

Gain(K::Parameter; inport::AbstractInPort = InPort(), outport::AbstractOutPort = OutPort())
Gain(;K::Parameter, inport::AbstractInPort = InPort(), outport::AbstractOutPort = OutPort())

A constructor for a Gain block

JuliaMBD._addinport! — Method

_addinport!(blk::BlockDefinition, p::AbstractInPort)

The function to add an inport

JuliaMBD._addoutport! — Method

_addoutport!(blk::BlockDefinition, p::AbstractOutPort)

The function to add an outport

JuliaMBD._compile_constructor — Method

compileconstructor(blk::BlockDefinition)

Create Expr to define the constructor

JuliaMBD._compile_function — Method

compilefunction(blk::BlockDefinition)

Create Expr to define the function

JuliaMBD.addblock! — Method

addblock!(blk::BlockDefinition, b::AbstractBlock)

The function to add a block

JuliaMBD.addparameter! — Method

addparameter!(blk::BlockDefinition, x::SymbolicValue) addparameter!(blk::BlockDefinition, x::SymbolicValue, val::Any)

The function to add a parameter.

JuliaMBD.expr — Method

expr(b)

A function to get Expr for main processes

JuliaMBD.expr_call — Method

expr_call(b)

A function to get Expr when the block b is called.

JuliaMBD.expr_defvalue — Method

expr_defvalue(x)

Create Expr to define the symbolic value.

JuliaMBD.expr_func — Method

expr_func

Expr for definition of function

JuliaMBD.expr_refvalue — Method

expr_refvalue(x)

Create Expr to refer the line.

JuliaMBD.expr_refvalue — Method

expr_refvalue(x)

Create Expr to refer the port.

JuliaMBD.expr_refvalue — Method

expr_refvalue(x)

Create Expr to refer the symbolic value.

JuliaMBD.expr_setvalue — Method

expr_setvalue(x, expr)

Create Expr to set x = expr.

JuliaMBD.expr_setvalue — Method

expr_setvalue(x, expr)

Create Expr to set x = expr.

JuliaMBD.expr_setvalue — Method

expr_setvalue(x, expr)

Create Expr to set x = expr.

JuliaMBD.get_default_inport — Method

get_default_inport(b)

A function to get a default inport. This function calls a specific getdefaultinport by a type of given block.

JuliaMBD.get_default_inport — Method

getdefaultinport(p::AbstractInPort) getdefaultoutport(p::AbstractInPort)

Get an instance of default in/out port.

JuliaMBD.get_default_inport — Method

getdefaultinport(p::AbstractOutPort) getdefaultoutport(p::AbstractOutPort)

Get an instance of default in/out port.

JuliaMBD.get_default_inport — Method

get_default_inport(b::UndefBlock)
get_default_outport(b::UndefBlock)

Return a nothing as an instance of default in/out port.

JuliaMBD.get_default_outport — Method

get_default_outport(b)

A function to get a default outport. This function calls a specific getdefaultoutport by a type of given block.

JuliaMBD.get_inports — Method

get_inports(b)
get_inports(b, key)

A function to get inports.

JuliaMBD.get_line — Method

get_line(x::AbstractInPort)

Get an input line of InPort.

JuliaMBD.get_lines — Method

get_lines(x::AbstractOutPort)

Get output lines from OutPort.

JuliaMBD.get_name — Method

get_name(x::AbstractPort)

Get a name (Symbol).

JuliaMBD.get_name — Method

get_name(x::SymbolicValue)

Get a name (Symbol).

JuliaMBD.get_outports — Method

get_outports(b)
get_outports(b, key)

A function to get outports.

JuliaMBD.get_params — Method

get_params(b)
get_params(b, key)

A function to get params.

JuliaMBD.get_parent — Method

get_parent(x::AbstractPort)

Get a parent block.

JuliaMBD.get_source — Method

get_source(x::AbstractLine)

Get a source of line.

JuliaMBD.get_target — Method

get_target(x::AbstractLine)

Get a target of line.

JuliaMBD.get_type — Method

get_type(b)

A function to get an instance of Val type

JuliaMBD.get_var — Method

get_var(x::AbstractLine)

Get a symbolicvar.

JuliaMBD.get_var — Method

get_var(x::AbstractPort)

Get a symbolicvar.

JuliaMBD.next — Method

next(b)

Get output blocks

JuliaMBD.prev — Method

prev(b)

Get input blocks

JuliaMBD.set_line! — Method

set_line!(x::AbstractInPort, l::AbstractLine)

Set an input line to InPort.

JuliaMBD.set_line! — Method

set_line!(x::AbstractOutPort, l::AbstractLine)

Set an output line to OutPort.

JuliaMBD.set_parent! — Method

set_parent!(x::AbstractPort, b::AbstractBlock)

Set a parent block.

JuliaMBD.tsort — Method

tsort

Tomprogical sort to determine the sequence of expression in SystemBlock